ensemble layers

+++ ucf101_ensemble/ensemble.py
@@ -156,7 +156,7 @@ if aggregate_method == 'average':
 
 elif aggregate_method == 'most_common':
   test_prob_max = tf.argmax(test_prob, axis=1)
-  test_class = tf.argmax(tf.bincount(test_prob_max))
+  test_class = tf.argmax(tf.bincount(test_prob_max), dtype=tf.int32)
 
 # verify if prediction is correct
 test_correct_pred = tf.equal(test_class, ph["y"])
